+++ changelog.md
@@ -1,3 +1,6 @@
+# Upcoming
+* Added config option to force the overtime regen overlay to always render on the first row of hearts, capping at 20 health. Useful for mods that change heart rendering
+
 # 2.0.2.0
 * Added config option to always show how much health will food regen instead of locking it behind advanced tooltips
 * Improved the overtime regen overlay texture. Thanks to @skippyzii